feat: implement football parser and scrapper with updated factory methods for sport-specific parsing
