+++ server/sports/services/parser.py
@@ -1,6 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Any
 
+
 class Parser(ABC):
     """
     Abstract base class for parsers.
@@ -34,9 +35,10 @@ class ParserFactory:
     """
 
     @staticmethod
-    def create_parser(parser_type) -> Parser:
-        if parser_type == "flashscore":
-            from sports.services.flashscore.flashscore_parser import FlashscoreParser
-            return FlashscoreParser()
+    def create_parser(parser_type: str, sport: str) -> Parser:
+        if parser_type.lower() == "flashscore" and sport.lower() == "football":
+            from sports.services.flashscore.football.football_parser import FootballParser
+
+            return FootballParser()
         else:
             raise ValueError(f"Unknown parser type: {parser_type}")
